f-footer@0.1.0 moved back footer.vue to components level and move tests to tests folder on components level
